+++ src/main/kotlin/dto/News.kt
@@ -1,3 +1,6 @@
+package dto
+
+import PlaceDeserializer
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import java.time.LocalDate
