+++ src/main/kotlin/dto/NewsApiResponse.kt
@@ -1,3 +1,5 @@
+package dto
+
 import kotlinx.serialization.Serializable
 
 @Serializable
