+++ src/main/resources/application.properties
@@ -1,0 +1,5 @@
+news.api.url=https://kudago.com/public-api/v1.4/news/
+news.api.defaultPageSize=100
+news.api.orderBy=-publication_date
+news.api.fields=id,title,place,description,site_url,favorites_count,comments_count,publication_date
+news.api.location=msk
