+++ src/main/resources/most_rated_news.csv
@@ -1,0 +1,4 @@
+id,title,place,description,siteUrl,favoritesCount,commentsCount,publicationDate
+51505,"История как на ладони: 3 канала в VK Видео, которые оживляют прошлое","","Изучение истории может быть по-настоящему захватывающим и интерактивным процессом. KudaGo и VK Видео собрали для вас интересные исторические видеоблоги, где можно найти разборы мифов и легенд, реконструкцию событий и анализ малоизвестных фактов, которые помогают по-новому взглянуть на различные исторические моменты.",https://kudago.com/all/news/istoriya-kak-na-ladoni-3-kanala/,1,0,2024-09-25
+51519,"Опубликован трейлер сериала Альфонсо Куарона с Кейт Бланшетт","","В проекте также участвуют Саша Барон Коэн и звезда «Игры в кальмара» Чон Хо Ён.",https://kudago.com/all/news/poyavilsya-trejler-seriala/,0,0,2024-09-25
+51518,"У Linkin Park вышла новая песня","","Она войдёт в грядущий альбом группы.",https://kudago.com/all/news/u-linkin-park-vyishla-novaya-pesnya/,0,0,2024-09-25
